requirements.txt
fastapi
uvicorn
pymysql
sqlalchemy
websockets
dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart